services/auth/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0 
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.7.1
pydantic-settings==2.10.1
python-dotenv==1.0.1
python-multipart

httpx==0.27.0
pytest==8.2.1
pytest-asyncio==0.23.6